refactor(apiType): 타입 네이밍 수정
