+++ src/api/apiType.ts
@@ -1,11 +1,11 @@
-interface CommonType {
+interface CommonMovieData {
   movieId: number;
   movieTitle: string;
   voteAverage: number;
   posterPath: string;
 }
 
-export interface MovieDataType extends CommonType {
+export interface MovieData extends CommonMovieData {
   voteCount: number;
   popularity: number;
   releaseDate: string;
@@ -13,7 +13,7 @@ export interface MovieDataType extends CommonType {
 
 type GenresType = { id: number; name: string }[];
 
-export interface MovieDetailDataType extends CommonType {
+export interface MovieDetailData extends CommonMovieData {
   genres: GenresType;
   overview: string;
 }
