+++ src/api/popularMovieList.ts
@@ -1,14 +1,14 @@
-import { MovieDataType } from './apiType';
+import { MovieData } from './apiType';
 import { catchError } from './errorStatus';
 import { wrappingMovieAPI } from './wrappingAPI';
 
 import { CONSTANT_URL, URL_LANGUAGE } from '../constant/api';
 
-export const getPopularMovieList = async (page = 1): Promise<MovieDataType[]> => {
+export const getPopularMovieList = async (page = 1): Promise<MovieData[]> => {
   const url = `${CONSTANT_URL.popularMovie}?${URL_LANGUAGE}&page=${page}`;
   const movieData = await catchError(url);
 
-  const dataCleaning = movieData.results.map((movieData: MovieDataType) => {
+  const dataCleaning = movieData.results.map((movieData: MovieData) => {
     return wrappingMovieAPI(movieData);
   });
